FU-218 refactor: group product form to reuse in product details page
